background-remover/requirements.txt
flatbuffers==23.3.3
humanfriendly==10.0
mpmath==1.3.0
numpy==1.23.5
onnxruntime==1.13.1
opencv-python-headless==4.6.0.66
packaging==23.1
Pillow==9.3.0
protobuf==4.22.3
rembg-aws-lambda==0.2.0
sympy==1.11.1